+++ RELEASES.md
@@ -8,6 +8,7 @@ Released on TBD (UTC).
 - Added missing serialization mappings for some instruments (#2702), thanks @faysou
 - Added support for DEX swaps for blockchain adapter (#2683), thanks @filipmacek
 - Added support for Pool liquidity updates for blockchain adapter (#2692), thanks @filipmacek
+- Added fill report reconciliation warning when discrepancy with existing fill (#2706), thanks @faysou
 
 ### Breaking Changes
 - Changed timer `allow_past=False` behavior: now validates the `next_event_time` instead of the `start_time`. This allows timers with past start times as long as their next scheduled event is still in the future